Split Cursor enterprise tests
